# Improve performance of `find_in_set` function