## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #.

## Rationale for this change
Improve performance of `find_in_set` function
<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

## What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

## Are these changes tested?
Yes. Added unit tests and benchmarks.
<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

## Are there any user-facing changes?
No.
<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->
*BENCHMARK RESULT
```rust
Compiling datafusion-functions v44.0.0 (/home/tailm/repos/github/datafusion/datafusion/functions)
    Finished `bench` profile [optimized] target(s) in 1m 02s
     Running benches/find_in_set.rs (target/release/deps/find_in_set-c80fb04778cbe610)
Gnuplot not found, using plotters backend
find_in_set/string_len_8
                        time:   [414.89 µs 415.67 µs 416.41 µs]
                        change: [-64.610% -64.526% -64.440%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 50 measurements (4.00%)
  2 (4.00%) low mild
find_in_set/string_view_len_8
                        time:   [440.65 µs 441.02 µs 441.50 µs]
                        change: [-63.746% -63.607% -63.494%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 50 measurements (4.00%)
  1 (2.00%) high mild
  1 (2.00%) high severe

find_in_set/string_len_32
                        time:   [474.24 µs 474.97 µs 475.79 µs]
                        change: [-61.751% -61.586% -61.447%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 50 measurements (4.00%)
  2 (4.00%) high mild
find_in_set/string_view_len_32
                        time:   [493.30 µs 494.79 µs 496.58 µs]
                        change: [-61.565% -61.365% -61.171%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 50 measurements (8.00%)
  1 (2.00%) high mild
  3 (6.00%) high severe

find_in_set/string_len_1024
                        time:   [4.8152 ms 4.8243 ms 4.8367 ms]
                        change: [-12.694% -12.465% -12.233%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 50 measurements (6.00%)
  2 (4.00%) high mild
  1 (2.00%) high severe
find_in_set/string_view_len_1024
                        time:   [4.8690 ms 4.8755 ms 4.8825 ms]
                        change: [-13.472% -13.274% -13.072%] (p = 0.00 < 0.05)
                        Performance has improved.
```