@tlm365 2025-01-08
@comphead Thanks for reviewing,

> I think it is a good PR the way it is. One thing comes to my mind which probably relevant for other string functions as well.
> 
> So we do `split(',')` for every search pattern value, but most likely that argument will be literal like `SELECT FIND_IN_SET(a, "s,q,l");` so we can do split only once.
> 
> I think we need to have a check a literal or column somewhere and we can optimize even further having this info, WDYT @tlm365 ?

Oh yes, your suggestion is very reasonable. I will try to implement it. TYSM ❤