@comphead 2025-01-08
Oh with literal support I think the code becomes much more tricky.
Wondering if the performance benefit still worthy such complications. @tlm365 can we a criterion to check `find_in_set` with literal vs original optimization?