@comphead 2025-01-10
> NOTE:
One notable thing I want to point out here. find_in_set(str, str_list) doesn't work if str or str_list is Scalar::Utf8View (string view literal), which is true for both the main branch and the original version of this PR (unit tests for these cases have been added in this PR, seems like a bug?! 🤔)

Is possible to provide a test case so we can file a ticket if this is undesired behavior?


The performance is definitely better, thanks @tlm365 I'm going to merge the PR